More consolitation of snowy platform. split fpga loader
