+++ Makefile
@@ -84,7 +84,7 @@ $(BUILD)/$(1)/%.o: %.s
 	@mkdir -p $$(dir $$@)
 	$(QUIET)$(CC) $(CFLAGS_$(1)) -c -o $$@ $$< 
 	
-$(BUILD)/$(1)/Resources/FPGA_%.o: Resources/FPGA_%.bin
+$(BUILD)/$(1)/Resources/%_fpga.o: Resources/%_fpga.bin
 	$(call SAY,[$(1)] FPGA $$<)
 	@mkdir -p $$(dir $$@)
 	$(QUIET)$(OBJCOPY) --rename-section .data=.rodata,contents,alloc,load,readonly,data -I binary -O elf32-littlearm -B armv5te $$< $$@
