+++ hw/platform/snowy/config.mk
@@ -4,6 +4,7 @@ CFLAGS_snowy += -DHSI_VALUE=16000000 -DREBBLE_PLATFORM=snowy -DREBBLE_PLATFORM_S
 
 SRCS_snowy = $(SRCS_snowy_family)
 SRCS_snowy += hw/platform/snowy/snowy.c
+SRCS_snowy += Resources/snowy_fpga.bin
 
 LDFLAGS_snowy = $(LDFLAGS_snowy_family)
 LIBS_snowy = $(LIBS_snowy_family)
