+++ hw/platform/snowy_family/config.mk
@@ -14,7 +14,7 @@ SRCS_snowy_family += hw/platform/snowy_family/snowy_scanlines.c
 SRCS_snowy_family += hw/platform/snowy_family/snowy_vibrate.c
 SRCS_snowy_family += hw/platform/snowy_family/snowy_ambient.c
 SRCS_snowy_family += hw/platform/snowy_family/snowy_ext_flash.c
-SRCS_snowy_family += Resources/FPGA_4.3_snowy_dumped.bin 
+SRCS_snowy_family += hw/platform/snowy_family/snowy_common.c
 
 LDFLAGS_snowy_family = $(LDFLAGS_stm32f4xx)
 LIBS_snowy_family = $(LIBS_stm32f4xx)
